requirements.txt
python3.6
cuda10.1
cudnn7.6
torch1.4
torchvision0.5
numpy
sacred
scipy
soundfile
kaldi_io
python_speech_features

